Refactor folder structure for ExpoApiResponse and import types
